refactor: split DatasetForm into folder with per-tab and field components
