+++ src/components/datasets/DatasetForm/RequiredMark.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const RequiredMark = () => <span style={{ color: "red" }}>*</span>;
+
+export default RequiredMark;
